:recycle: :lipstick:  Refactored app + Updated Menu
