+++ Assets/BVH/pointeur_2.bvh
