+++ Assets/BVH/pointeur_3.bvh
