--- Assets/Scripts/BvhImporter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1154d8de4ec81974e89c800cc388bc52
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
