+++ Assets/Scripts/MenuSystem/MenuManager.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
 public class MenuManager : MonoBehaviour
